chore: move Configuration to single feature file
