+++ packages/core/admin/admin/src/components/UnauthenticatedLogo.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { useConfiguration } from '../contexts/configuration';
+import { useConfiguration } from '../features/Configuration';
 
 const Img = styled.img`
   height: ${72 / 16}rem;
@@ -9,7 +9,7 @@ const Img = styled.img`
 const Logo = () => {
   const {
     logos: { auth },
-  } = useConfiguration();
+  } = useConfiguration('UnauthenticatedLogo');
 
   return <Img src={auth?.custom?.url ?? auth.default} aria-hidden alt="" />;
 };
