+++ packages/core/admin/admin/tests/utils.tsx
@@ -31,7 +31,7 @@ import { reducer as cmAppReducer } from '../src/content-manager/pages/App';
 import { reducer as editViewReducer } from '../src/content-manager/pages/EditViewLayoutManager';
 import { reducer as listViewReducer } from '../src/content-manager/pages/ListViewLayoutManager';
 import { reducer as crudReducer } from '../src/content-manager/sharedReducers/crud/reducer';
-import { ConfigurationContextProvider } from '../src/contexts/configuration';
+import { _internalConfigurationContextProvider as ConfigurationContextProvider } from '../src/features/Configuration';
 import { reducer as appReducer } from '../src/reducer';
 
 import { server } from './server';
@@ -110,11 +110,11 @@ const Providers = ({ children, initialEntries }: ProvidersProps) => {
                       <ConfigurationContextProvider
                         showReleaseNotification={false}
                         showTutorials={false}
-                        updateProjectSettings={jest.fn()}
                         logos={{
                           auth: { default: '' },
                           menu: { default: '' },
                         }}
+                        updateProjectSettings={jest.fn()}
                       >
                         {children}
                       </ConfigurationContextProvider>
